Fixes refactored imports
