+++ types.tsx
@@ -1,7 +1,7 @@
 import { StackNavigationProp } from '@react-navigation/stack';
 import { StringValueNode } from 'graphql/language/ast';
 import { LatLng } from 'react-native-maps';
-import { JobFilter } from './features/job/JobList';
+import { JobFilter } from './components/FilterPills';
 
 export type RootStackParamList = {
     Home: undefined;
